Add automatic micromanager discovery for linux
If you follow the steps in the Docker file for pymmcore
then you end up with an installation in /usr/local/lib/micro-manager/

Also, paths are case senstive on linux hence so search for both capital
and lowercase `m` in the glob.

Finally, added the searched path to the logger if nothing was found in
case people have non-standard install locations.